stabilize TFFMClassifier
