+++ tffm/models.py
@@ -7,8 +7,9 @@ import shutil
 from tqdm import tqdm
 
 
-from .core import TFFMCore, loss_logistic, loss_mse
+from .core import TFFMCore
 from .base import TFFMBaseModel
+from .utils import loss_logistic, loss_mse
 
 
 
@@ -18,101 +19,9 @@ class TFFMClassifier(TFFMBaseModel):
     This class implements L2-regularized arbitrary order FM model with logistic
     loss and gradient-based optimization.
 
-    It supports arbitrary order of interactions and has linear complexity in the
-    number of features (a generalization of the approach described in Lemma 3.1
-    in the referenced paper, details will be added soon).
-
-    It can handle both dense and sparse input. Only numpy.array and CSR matrix are
-    allowed as inputs; any other input format should be explicitly converted.
-
-    Support logging/visualization with TensorBoard.
-
     Only binary classification with 0/1 labels supported.
 
-    Parameters
-    ----------
-    rank : int
-        Number of factors in low-rank appoximation.
-        This value is shared across different orders of interaction.
-
-    order : int, default: 2
-        Order of corresponding polynomial model.
-        All interaction from bias and linear to order will be included.
-
-    optimizer : tf.train.Optimizer, default: AdamOptimizer(learning_rate=0.1)
-        Optimization method used for training
-
-    batch_size : int, default: -1
-        Number of samples in mini-batches. Shuffled every epoch.
-        Use -1 for full gradient (whole training set in each batch).
-
-    n_epoch : int, default: 100
-        Default number of epoches.
-        It can be overrived by explicitly provided value in fit() method.
-
-    reg : float, default: 0
-        Strength of L2 regularization
-
-    init_std : float, default: 0.01
-        Amplitude of random initialization
-
-    input_type : str, 'dense' or 'sparse', default: 'dense'
-        Type of input data. Only numpy.array allowed for 'dense' and
-        scipy.sparse.csr_matrix for 'sparse'. This affects construction of
-        computational graph and cannot be changed during training/testing.
-
-    log_dir : str or None, default: None
-        Path for storing model stats during training. Used only if is not None.
-        WARNING: If such directory already exists, it will be removed!
-        You can use TensorBoard to visualize the stats:
-        `tensorboard --logdir={log_dir}`
-
-    session_config : tf.ConfigProto or None, default: None
-        Additional setting passed to tf.Session object.
-        Useful for CPU/GPU switching.
-        `tf.ConfigProto(device_count = {'GPU': 0})` will disable GPU (if enabled)
-
-
-    verbose : int, default: 0
-        Level of verbosity.
-        Set 1 for tensorboard info only and 2 for additional stats every epoch.
-
-    Attributes
-    ----------
-    core : TFFMCore or None
-        Computational graph with internal utils.
-        Will be initialized during first call .fit()
-
-    session : tf.Session or None
-        Current execution session or None.
-        Should be explicitly terminated via calling destroy() method.
-
-    steps : int
-        Counter of passed lerning epochs, used as step number for writing stats
-
-    n_features : int
-        Number of features used in this dataset.
-        Inferred during the first call of fit() method.
-
-    intercept : float, shape: [1]
-        Intercept (bias) term.
-
-    weights : array of np.array, shape: [order]
-        Array of underlying representations.
-        First element will have shape [n_features, 1],
-        all the others -- [n_features, rank].
-
-    Notes
-    -----
-    You should explicitly call destroy() method to release resources.
-    Parameter rank is shared across all orders of interactions (except bias and
-    linear parts).
-    tf.sparse_reorder doesn't requied since COO format is lexigraphical ordered.
-
-    References
-    ----------
-    Steffen Rendle, Factorization Machines
-        http://www.csie.ntu.edu.tw/~b97053/paper/Rendle2010FM.pdf
+    See TFFMBaseModel docs for details about parameters.
     """
 
     def __init__(self, rank=2, order=2, input_type='dense', n_epochs=100,
@@ -135,7 +44,7 @@ class TFFMClassifier(TFFMBaseModel):
 
     def preprocess_target(self, y_):
         # suppose input {0, 1}, but use instead {-1, 1} labels
-        assert(set(y_)==set([0, 1]))
+        assert(set(y_) == set([0, 1]))
         return y_ * 2 - 1
 
     def predict(self, X):
@@ -153,7 +62,6 @@ class TFFMClassifier(TFFMBaseModel):
         """
         raw_output = self.decision_function(X)
         predictions = (raw_output > 0).astype(int)
-        # TODO: preprocess classes
         return predictions
 
     def predict_proba(self, X):
